python/requirements.txt
flask==3.0.3
pillow==10.4.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
psycopg2-binary==2.9.9


